==> 01a - process wands dataset.ipynb <==
import pandas as pd
import wands
from tqdm import tqdm
import os

WANDS_REPO = '/usr/local/google/home/raulramos/projects/WANDS'


pw = pd.read_csv(f'{WANDS_REPO}/dataset/product.csv', sep='\t')
pw.set_index('product_id', inplace=True)

qw = pd.read_csv(f'{WANDS_REPO}/dataset/query.csv', sep='\t')
qw.set_index('query_id', inplace=True)

lw = pd.read_csv(f'{WANDS_REPO}/dataset/label.csv', sep='\t')



# queries already come in cool format
qw.to_parquet(f'wands/queries.parquet')
qw.head()

p = pw
p = [ [product_id, utils.get_wands_product_string(row)] for product_id, row in tqdm(pw.iterrows(), total=len(p))]
p = pd.DataFrame(p, columns=['product_id', 'product']).set_index('product_id')
p.to_parquet(f'wands/products.parquet')
p.head()

lw['label'].value_counts()

m = {'Exact': 2, 'Partial': 1, 'Irrelevant': 0}`
lw['relevance'] = [m[i] for i in lw['label'].values]

lw.head()

lw.to_parquet('wands/relevance.parquet')
