==> 01b - process esci dataset.ipynb <==
import pandas as pd
import utils
from tqdm import tqdm
import os

ESCI_DATASET_ROOT = '/usr/local/google/home/raulramos/projects/esci-data'
locale = 'es'
dataset = f'esci-{locale}'

dgt = utils.load_esci_examples(ESCI_DATASET_ROOT=ESCI_DATASET_ROOT, locale=locale)

queries = dgt.groupby('query_id').first()[['query']]

products = dgt.groupby('product_id').first()[['product_title', 'product_description', 'product_bullet_point', 'product_brand', 'product_color']]
products = pd.DataFrame([{product_id: utils.get_esci_product_string(pi) for product_id,pi in tqdm(products.iterrows())}], index=['product']).T

queries.shape, products.shape

queries.to_parquet(f'{dataset}/queries.parquet')
products.to_parquet(f'{dataset}/products.parquet')
dgt[['example_id', 'query_id', 'product_id', 'esci_label', 'relevance']].to_parquet(f'{dataset}/relevance.parquet')
