==> 02a - gemini - generate embeddings.ipynb <==
import pandas as pd
import wands
from tqdm import tqdm
import os
import sys
sys.path.append('..')

import utils

dataset = 'wands'
model_folder = 'text-exp0307'
model_id = 'text-embedding-large-exp-03-07'

model_folder = 'text-004'
model_id = 'text-embedding-004'

q = pd.read_parquet(f'{dataset}/queries.parquet')
p = pd.read_parquet(f'{dataset}/products.parquet')


project_id = 'gemma-test-deployment'

q.head()

queries = q
queries['embeddings']= list(utils.get_gemini_embeddings(list(queries['query'].values), 
                                                   model=model_id,
                                                   project_id = project_id,
                                                   location = "us-central1",
                                                   task_type = 'RETRIEVAL_QUERY', 
                                                  n_jobs=-1))


queries.to_parquet(f'{model_folder}/{dataset}-queries.parquet')

queries.head()

products = p

fname = f'{model_folder}/{dataset}-products.parquet'

split_size = 10000
for i in range(0,len(products),split_size):
    split_fname = f'{fname}-split-{i:07d}'
    product_split = products.iloc[i:i+split_size].copy()
    print (split_fname, product_split.shape, flush=True)
    if os.path.isfile(split_fname):
        continue
    product_split['embeddings']= list(utils.get_gemini_embeddings(list(product_split['product'].values), 
                                                   model=model_id,
                                                   project_id = project_id,
                                                   location = "us-central1",
                                                   task_type = 'RETRIEVAL_DOCUMENT', 
                                                  n_jobs=-1))

    product_split.to_parquet(split_fname)

## collate files

from glob import glob
import pandas as pd

sorted(glob(f'{model_folder}/{dataset}-products.parquet-split-*'))

all_products = pd.concat([pd.read_parquet(i) for i in sorted(glob(f'{model_folder}/{dataset}-products.parquet-split-*'))])

all_products.shape

all_products.to_parquet(f'{model_folder}/{dataset}-products.parquet')
