==> 02b - openai - generate embeddings.ipynb <==
import pandas as pd
import wands
from tqdm import tqdm
import os
import sys
sys.path.append('..')

import utils

with open('../../../secrets/openai-key.txt') as f:
    OPENAI_API_KEY = f.read().strip()


dataset = 'wands'
model_folder = 'openai'
model_id = 'text-embedding-large-exp-03-07'

# load products and queries


q = pd.read_parquet(f'{wands}/queries.parquet')
p = pd.read_parquet(f'{wands}/products.parquet')

q.shape, p.shape

# generate embeddings

queries = q
queries['embeddings'] = utils.get_openai_batch_embeddings(queries['query'].values, OPENAI_API_KEY, chunk_size=500)

queries.to_parquet(f'{model_folder}/{dataset}-queries.parquet')

products = p
products['embeddings'] = utils.get_openai_batch_embeddings(products['product'].values, OPENAI_API_KEY, chunk_size=500)

products.to_parquet(f'{model_folder}/{dataset}-products.parquet')

products
