==> 03 - rank queries.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
from joblib import Parallel, delayed
import polars as pl
import os
import utils

# select model and dataset

model_folder = 'openai'
#model_folder = 'text-exp0307'
#model_folder = 'text-004'

#dataset = 'wands'
dataset = 'esci-us'


# load embeddings and ground truth

q = pd.read_parquet(f'{model_folder}/{dataset}-queries.parquet')

p = pl.read_parquet(f'{model_folder}/{dataset}-products.parquet').to_pandas()
p.set_index('__index_level_0__', inplace=True)

r = pd.read_parquet(f'{dataset}/relevance.parquet')

# index ground truth with query and product ids
r.index = [f'{row.query_id}|{row.product_id}' for _,row in tqdm(r.iterrows(), total=len(r))]

q.shape, p.shape, r.shape

r.head()

q.head()

p.head()

# rank topn according dot product on embeddings

qe = np.stack(q.embeddings.values.copy())
pe = np.stack(p.embeddings.values.copy())

qe.shape, pe.shape

## vector search using `faiss`

import faiss

# search in normalized vector space is equivalent to dot product
kqe = qe / np.linalg.norm(qe, axis=1).reshape(-1,1)
kpe = pe / np.linalg.norm(pe, axis=1).reshape(-1,1)

index = faiss.IndexFlatL2(kqe.shape[1])   # build the index
index.add(kpe)
print(index.is_trained, index.ntotal)

chunk_size = 2048
qnn = []
for i in tqdm(range(0,len(qe), chunk_size)):
    _,idxs = index.search(kqe[i:i+chunk_size], 40) # sanity check
    for ii in idxs:
        qnn.append(ii)

q['topn'] = [list(p.index[qnni]) for qnni in tqdm(qnn)]

## brute force approach with numpy dotproduct and `numpy.argsort`


# brute force approach with numpy dotproduct and argsort

def get_dotp_closests_idxs(q_embedding,p_embeddings_matrix, topn=40):
    return np.argsort( p_embeddings_matrix.dot(q_embedding) )[-topn:][::-1]

#qnn = Parallel(n_jobs=-1, verbose=5)(delayed(get_dotp_closests_idxs)(qi,pe, topn=40) for qi in qe)
#q['topn'] = [list(p.index[qnni]) for qnni in tqdm(qnn)]

1

# create rankings file together with ground truth

qr = []
for query_id, qi in tqdm(q.iterrows()):
    for i, product_id in enumerate(qi.topn):
        qr.append([query_id, product_id, i+1])
qr = pd.DataFrame(qr, columns = ['query_id', 'product_id', 'model_rank'])
qr.index = [f'{row.query_id}|{row.product_id}' for _,row in tqdm(qr.iterrows(), total=len(qr))]
qr = pd.merge(qr, r[['relevance']], how='left', left_index=True, right_index=True).fillna(0)
qr.shape

qr.head()

qr.relevance.value_counts()

qr.to_parquet(f'{model_folder}/{dataset}-ranking_with_relevance.parquet', index=False)

qr[qr.product_id=='B000MOO21W']

q
