==> 04 - compute ndcg.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
from joblib import Parallel, delayed
from sklearn.metrics import ndcg_score
import utils
import polars as pl

# compute metrics for all datasets and models

def compute_metrics(model_folder, dataset, relevance='binary_strict'):
    r = pd.read_parquet(f'{model_folder}/{dataset}-ranking_with_relevance.parquet')

    if relevance=='binary_strict':
        if dataset == 'wands':
            r.relevance=[1 if i==2 else 0 for i in r.relevance]
        elif dataset.startswith('esci'):
            r.relevance=[1 if i==4 else 0 for i in r.relevance]

    elif relevance=='binary_tolerant':
        if dataset == 'wands':
            r.relevance=[1 if i>0 else 0 for i in r.relevance]
        elif dataset.startswith('esci'):
            r.relevance=[1 if i>1 else 0 for i in r.relevance]
    
    elif relevance!='asis':
        raise ValueError(f"invalid relevance string {relevance}'")
    
    df_dcg, ndcg = utils.compute_ndcg(r)

    custom_ndcg = ndcg.ndcg.mean()

    sklearn_ndcg = []
    for qid in tqdm(np.unique(r.query_id)):
        rq = r[r.query_id==qid]
        sklearn_ndcg.append(ndcg_score(rq.relevance.values.reshape(1,-1), 1/rq.model_rank.values.reshape(1,-1)))

    sklearn_ndcg = np.mean(sklearn_ndcg)

    return {'custom_ndcg': custom_ndcg, 'sklearn_ndcg': sklearn_ndcg}

## relevance strict

r = {}

for model_folder in ['openai', 'text-exp0307', 'text-004']:
    r[model_folder] = {}
    for dataset in ['wands', 'esci-es', 'esci-us']:
        print(model_folder, dataset, flush=True)
        r[model_folder][dataset] = compute_metrics(model_folder, dataset)

1

pd.DataFrame({m: {d: r[m][d]['custom_ndcg'] for d in r[m].keys()} for m in r.keys()}).T

pd.DataFrame({m: {d: r[m][d]['sklearn_ndcg'] for d in r[m].keys()} for m in r.keys()}).T

# relevance tolerant

r = {}

for model_folder in ['openai', 'text-exp0307', 'text-004']:
    r[model_folder] = {}
    for dataset in ['wands', 'esci-es', 'esci-us']:
        print(model_folder, dataset, flush=True)
        r[model_folder][dataset] = compute_metrics(model_folder, dataset, relevance='binary_tolerant')

pd.DataFrame({m: {d: r[m][d]['custom_ndcg'] for d in r[m].keys()} for m in r.keys()})

pd.DataFrame({m: {d: r[m][d]['sklearn_ndcg'] for d in r[m].keys()} for m in r.keys()})

# relevance as is

r = {}

for model_folder in ['openai', 'text-exp0307', 'text-004']:
    r[model_folder] = {}
    for dataset in ['wands', 'esci-es']:
        print(model_folder, dataset, flush=True)
        r[model_folder][dataset] = compute_metrics(model_folder, dataset, relevance='asis')

pd.DataFrame({m: {d: r[m][d]['custom_ndcg'] for d in r[m].keys()} for m in r.keys()})

pd.DataFrame({m: {d: r[m][d]['sklearn_ndcg'] for d in r[m].keys()} for m in r.keys()})

# summarize datasets

import polars as pl

datasets = ['wands', 'esci-es', 'esci-us']

stats = []

for dataset in datasets:
    print (dataset, flush=True)
    p = pl.read_parquet(f'{dataset}/products.parquet').to_pandas()
    q = pd.read_parquet(f'{dataset}/queries.parquet')
    r = pd.read_parquet(f'{dataset}/relevance.parquet')

    s = {'num_products': len(p), 
         'num_queries': len(q), 
         'num_relevance_judgements': len(r),
         'num_queries_with_relevance': len(np.unique(r.query_id)),
         'num_products_with_relevance': len(np.unique(r.product_id)),
         'relevance_values_distribution': r.relevance.value_counts().sort_index(ascending=False) / len(r),
         'mean_num_relevance_judgements_per_query': r.groupby('query_id').size().values.mean()
        }
    stats.append(s)

stats = pd.DataFrame(stats, index=datasets).T

stats

for dataset in datasets:
    print (stats.loc['relevance_values_distribution'][dataset])

# manual calculation for inspection

#model_folder = 'openai'
#model_folder = 'text-exp0307'
model_folder = 'text-004'
#dataset = 'wands'
dataset = 'esci-us'

r = pd.read_parquet(f'{model_folder}/{dataset}-ranking_with_relevance.parquet')
r.head()    

len(np.unique(r.query_id)), len(np.unique(r.product_id))

p = pl.read_parquet(f'{dataset}/products.parquet').to_pandas()
p.set_index('__index_level_0__', inplace=True)
p.head()

p.shape

if dataset == 'wands':
    r.relevance=[1 if i==2 else 0 for i in r.relevance]
elif dataset.startswith('esci'):
    r.relevance=[1 if i==4 else 0 for i in r.relevance]

r.relevance.value_counts()

## custom ndcg

df_dcg, ndcg = utils.compute_ndcg(r)

ndcg.ndcg.mean()

## sklearn ndcg

sklearn_ndcg = []
for qid in tqdm(np.unique(r.query_id)):
    rq = r[r.query_id==qid]
    sklearn_ndcg.append(ndcg_score(rq.relevance.values.reshape(1,-1), 1/rq.model_rank.values.reshape(1,-1)))

np.mean(sklearn_ndcg)

ndcg.ndcg.mean()

# custom ndcg

```
                 wands           esci-es          esci-es        esci-us
               relevance-01   relevance-01234   relevance-01    relevance-01
text-004          0.839           0.666            0.654          0.671
text-ext307       0.859           0.751            0.751          0.705          
openai            0.825           0.739            0.724
```

## skelarn ndcg

```
                 wands           esci-es          esci-es
               relevance-01   relevance-01234   relevance-01
text-004          0.619           0.586              0.554         0.606       
text-ext307       0.638           0.724              0.696
openai            0.615           0.705              0.676
```

pd.DataFrame({m: {d: r[m][d]['custom_ndcg'] for d in r[m].keys()} for m in r.keys()}).T
