==> 06 - test nDCG.ipynb <==
import utils
import pandas as pd
from tqdm import tqdm
import numpy as np

## according to [Wikipedia nDCG](https://en.wikipedia.org/wiki/Discounted_cumulative_gain) and  [this blog](https://medium.com/data-science/demystifying-ndcg-bee3be58cfe0)


$$nDCG = \frac{\sum_{m \in \texttt{model\_ranking}} \frac{relevance_m}{\log (m+1)}}{\sum_{i \in \texttt{ideal\_ranking}} \frac{relevance_i}{\log (i+1)}}$$


r = pd.read_parquet('text-004/esci-es-ranking_with_relevance.parquet')
r

r = pd.DataFrame([ [1,1,1,0.0],
               [1,2,2,0.0],
               [1,3,3,1.0],
               [1,4,4,1.0],
               [1,5,5,1.0],
               [2,1,1,1.0],
               [2,2,2,0.0],
               [2,3,3,1.0],
               [2,4,4,0.0],
               [2,5,5,1.0],
               [3,1,1,3.0],
               [3,2,2,2.0],
               [3,3,3,3.0],
               [3,4,4,0.0],
               [3,5,5,1.0],
               [3,6,6,2.0]                 
                 
                 ] , 
             columns = ['query_id', 'product_id', 'model_rank', 'relevance'])

r

qdcg, ndcg = utils.compute_ndcg(r)

ndcg

for qi in np.unique(r.query_id):

ndcg.ndcg.mean()

# Sklearn implementation (only order maters)

from sklearn.metrics import ndcg_score
import numpy as np

# we have ground-truth relevance of some answers to a query:
true_relevance = np.asarray([[10, 0, 0, 1, 5]])
# we predict some scores (relevance) for the answers
scores        = np.asarray([[.1, .2, .3, 4, 70]])
ndcg_score(true_relevance, scores)

scores        = np.asarray([[10, 20, 39, 40, 70]])
ndcg_score(true_relevance, scores)

true_relevance.dtype, scores.dtype, true_relevance.shape, scores.shape

true_r.shape, scores_r.shape

true_r = r[r.query_id==1].relevance.values.astype(int).reshape(1,-1)
scores_r = 1/r[r.query_id==1].model_rank.values.reshape(1,-1)
ndcg_score(true_r, scores_r)

# we have ground-truth relevance of some answers to a query:
true_relevance = np.asarray([[10, 0, 0, 1, 5]])
# we predict some scores (relevance) for the answers
scores        = np.asarray([[10, .2, .3, 1, 2]])
ndcg_score(true_relevance, scores)
